- Single Step forms will have to use ###TEMPLATE_FORM1### and ###submit_nextStep###
- Removed Controller Dropdown in Flexform.
- Manual Update

New Feature: Send CC-Emails

git-svn-id: https://svn.typo3.org/TYPO3v4/Extensions/mailformplusplus/trunk@19215 735d13b6-9817-0410-8766-e36946ffe9aa

+++ Classes/Finisher/F3_MailformPlusPlus_Finisher_StoreUploadedFiles.php
@@ -83,17 +83,18 @@ class F3_MailformPlusPlus_Finisher_StoreUploadedFiles extends F3_MailformPlusPlu
 		if(isset($_SESSION['mailformplusplusFiles']) && is_array($_SESSION['mailformplusplusFiles']) && strlen($newFolder) > 0 ) {
 			foreach($_SESSION['mailformplusplusFiles'] as $field=>$files) {
 				foreach($files as $key => $file) {
-					$newFilename = $this->getNewFilename($file['uploaded_name']);
-
-					F3_MailformPlusPlus_StaticFuncs::debugMessage('copy_file',$file['uploaded_path'].$file['uploaded_name'],$uploadPath.$newFilename);
-					copy($file['uploaded_path'].$file['uploaded_name'],$uploadPath.$newFilename);
-					unlink($file['uploaded_path'].$file['uploaded_name']);
-
-					$_SESSION['mailformplusplusFiles'][$field][$key]['uploaded_path'] = $uploadPath;
-					$_SESSION['mailformplusplusFiles'][$field][$key]['uploaded_name'] = $newFilename;
-					$_SESSION['mailformplusplusFiles'][$field][$key]['uploaded_folder'] = $newFolder;
-					$_SESSION['mailformplusplusFiles'][$field][$key]['uploaded_url'] = t3lib_div::getIndpEnv('TYPO3_SITE_URL').$newFolder.$newFilename;
-
+					if($file['uploaded_path'] != $uploadPath) {
+						$newFilename = $this->getNewFilename($file['uploaded_name']);
+	
+						F3_MailformPlusPlus_StaticFuncs::debugMessage('copy_file',$file['uploaded_path'].$file['uploaded_name'],$uploadPath.$newFilename);
+						copy($file['uploaded_path'].$file['uploaded_name'],$uploadPath.$newFilename);
+						unlink($file['uploaded_path'].$file['uploaded_name']);
+	
+						$_SESSION['mailformplusplusFiles'][$field][$key]['uploaded_path'] = $uploadPath;
+						$_SESSION['mailformplusplusFiles'][$field][$key]['uploaded_name'] = $newFilename;
+						$_SESSION['mailformplusplusFiles'][$field][$key]['uploaded_folder'] = $newFolder;
+						$_SESSION['mailformplusplusFiles'][$field][$key]['uploaded_url'] = t3lib_div::getIndpEnv('TYPO3_SITE_URL').$newFolder.$newFilename;
+					}
 				}
 			}
 		}
